contract_connector/requirements.txt
Flask==3.1.0
web3==7.10.0
python-dotenv==1.1.0
pandas==2.2.3